requirements.txt
discord
apscheduler
num2words
psutil
requests
urbandict
googletrans
pillow
path
py-dactyl
wisp